Support Libraries added
Minor tweak to allow for an enhanced command line functionality and
support. Allows for a clean help menu, use of options, colorization,
etc ...

Included a library for handling persistence of data and dynamic query
of the data with type-safety checks.

The intent is to allow ruby scripts used for monitoring to include a
single monitor.rb allowing for a framework of functionality in a DSL
like language for easy monitoring.