Added Library Support for URL Monitoring
Updated the project to dynamically append the libraries into the path
for executed ruby scripts. This allows ruby based monitors to have
access to a suite of libraries for making monitors easier to use.

Added a 'monitor' library so that some url monitoring can be made a bit
easier and be more readable/maintainable.