URL Check supports Location response header
Fixed crazy attempt at ruby 1.9 support when it was not broken to
start with. The ./ prefix caused it to look for the required library
in a different location then the RUBYLIB path.

URL Checker now detects if a location header exists and stores that for
use as the location that the next url check should go to.

Check.url now returns the entire UrlCheck object instead of only the
cookies portion of that object. Made better logical sense.