URL Check Enhancements for Ease of Use
Moved the require net/https to the initialize to prevent needing to
install gems for things that are not going to be used. This is in
preparation for other monitor types like through ssh, pop3, etc ...

Only store the Location redirect information if the status code is of
a redirect type (300 - 399 status codes).

Cleaned up the internal checks to not access the variables directly
using the class variable notation but rather through the accessor
methods.

Added a Check.url_redirect checker that takes an existing UrlCheck
object and will automatically goto the location specified in the
redirect and pass along any cookies and refuse POST parameters. This
gives a much simplier interface for handling url redirects and makes
it much more accurate to the behavior of a real browser.