Improved messaging to be more meaningful
Only compile regex pattern for cookie parsing if cookies are returned.

Error messages for connection failures indicate url that was tried now.

Fixed an issue with check using accessor for error count because check
was also being called outside of the block scope by the initialization
routine. Cleanest solution was to simply let error count be direct for
now. This might cause an issue with subclasses of UrlCheck however but
will deal with that later if needed.

Standard messaging for various checks are more meaningful in telling
what the actual value was if it was not the expected value.