URL Checks support GET
If no parameters are given, then the url check invoked will be a GET
instead of a POST.

Fixed issue where query parameters were not being included in the URL
check.

Added check_redirect support to make a standardized way of checking
that the redirects are correct with dynamic informative messaging when
it is not.